feat: - added permissions models, schemas, repository & handler - added - added methods for fetching data (select) from db to base crudmixin - updated endpoints and added rights check - updated routes and added organization_id query param due to necessity of rights check - added authorization with bearer token possibility to openapi docs - added exception handler to app

+++ backend/dataforce_studio/api/orbits/orbits_members.py
@@ -1,36 +1,71 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, Request
 
 from dataforce_studio.handlers.orbits import OrbitHandler
+from dataforce_studio.handlers.permissions import PermissionsHandler
 from dataforce_studio.schemas.orbit import (
     OrbitMember,
     OrbitMemberCreate,
     UpdateOrbitMember,
 )
+from dataforce_studio.schemas.permissions import Resources, ResourceAction
 
 orbit_members_router = APIRouter(
-    prefix="/orbits/{orbit_id}/members", tags=["orbits-members"]
+    prefix="/{organization_id}/orbits/{orbit_id}/members", tags=["orbits-members"]
 )
 
 orbit_handler = OrbitHandler()
+permissions_handler = PermissionsHandler()
 
 
 @orbit_members_router.get("", response_model=list[OrbitMember])
 async def get_orbit_members(
+    request: Request,
+    organization_id: int,
     orbit_id: int,
 ) -> list[OrbitMember]:
+    await permissions_handler.check_organization_permission(
+        organization_id,
+        request.user.id,
+        Resources.ORBIT_USER,
+        ResourceAction.LIST,
+    )
     return await orbit_handler.get_orbit_members(orbit_id)
 
 
 @orbit_members_router.post("", response_model=OrbitMember)
-async def add_member_to_orbit(member: OrbitMemberCreate) -> OrbitMember:
+async def add_member_to_orbit(
+    request: Request, organization_id: int, member: OrbitMemberCreate
+) -> OrbitMember:
+    await permissions_handler.check_organization_permission(
+        organization_id,
+        request.user.id,
+        Resources.ORBIT_USER,
+        ResourceAction.CREATE,
+    )
     return await orbit_handler.create_orbit_member(member)
 
 
 @orbit_members_router.patch("/{member_id}", response_model=OrbitMember)
-async def update_orbit_member(member: UpdateOrbitMember) -> OrbitMember:
+async def update_orbit_member(
+    request: Request, organization_id: int, member: UpdateOrbitMember
+) -> OrbitMember:
+    await permissions_handler.check_organization_permission(
+        organization_id,
+        request.user.id,
+        Resources.ORBIT_USER,
+        ResourceAction.UPDATE,
+    )
     return await orbit_handler.update_orbit_member(member)
 
 
 @orbit_members_router.delete("/{member_id}", status_code=204)
-async def remove_orbit_member(member_id: int) -> None:
+async def remove_orbit_member(
+    request: Request, organization_id: int, member_id: int
+) -> None:
+    await permissions_handler.check_organization_permission(
+        organization_id,
+        request.user.id,
+        Resources.ORBIT_USER,
+        ResourceAction.DELETE,
+    )
     return await orbit_handler.delete_orbit_member(member_id)
